Exception update

+++ src/main/java/bookstore/admin/panel/service/AuthorService.java
@@ -1,8 +1,10 @@
 package bookstore.admin.panel.service;
 
 import bookstore.admin.panel.dao.entity.Author;
+import bookstore.admin.panel.dao.entity.Book;
 import bookstore.admin.panel.dao.repository.AuthorRepository;
 import bookstore.admin.panel.dao.repository.BookRepository;
+import bookstore.admin.panel.exception.BadRequestException;
 import bookstore.admin.panel.exception.Error;
 import bookstore.admin.panel.exception.NotFoundException;
 import bookstore.admin.panel.mapper.UniversalMapper;
@@ -12,6 +14,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -26,30 +29,47 @@ public class AuthorService {
     }
 
     public AuthorGetDto getAuthorById(Long id) {
+        if (Objects.isNull(id)) {
+            throw new BadRequestException(Error.BAD_REQUEST_ERROR_CODE, Error.BAD_REQUEST_ERROR_MESSAGE);
+        }
         Author foundAuthor = authorRepository.findById(id).orElseThrow(()
-                -> new NotFoundException(Error.BOOK_NOT_FOUND_ERROR_CODE,
-                Error.BOOK_NOT_FOUND_ERROR_MESSAGE));
+                -> new NotFoundException(Error.NOT_FOUND_ERROR_CODE,
+                Error.AUTHOR_NOT_FOUND_ERROR_MESSAGE));
         return mapper.toAuthorGetDto(foundAuthor);
     }
 
     public void addAuthor(AuthorDto authorDto) {
+        if (Objects.isNull(authorDto)) {
+            throw new BadRequestException(Error.BAD_REQUEST_ERROR_CODE, Error.BAD_REQUEST_ERROR_MESSAGE);
+        }
         Author author = mapper.toAuthorEntity(authorDto);
-        author.setBooks(bookRepository.findAllByIdIn(authorDto.getBookIdList()));
-        System.out.println(author.getBooks());
+        if (Objects.isNull(author.getName())) {
+            throw new BadRequestException(Error.BAD_REQUEST_ERROR_CODE, Error.BAD_REQUEST_ERROR_MESSAGE);
+        }
+        List<Book> books = bookRepository.findAllByIdIn(authorDto.getBookIdList());
+        if (Objects.nonNull(books)) {
+            author.setBooks(books);
+        }
         authorRepository.save(author);
     }
 
     public void deleteAuthorById(Long id) {
+        if (Objects.isNull(id)) {
+            throw new BadRequestException(Error.BAD_REQUEST_ERROR_CODE, Error.BAD_REQUEST_ERROR_MESSAGE);
+        }
         authorRepository.findById(id).orElseThrow(
-                ()-> new NotFoundException(Error.BOOK_NOT_FOUND_ERROR_CODE,
-                Error.BOOK_NOT_FOUND_ERROR_MESSAGE));
+                () -> new NotFoundException(Error.NOT_FOUND_ERROR_CODE,
+                        Error.AUTHOR_NOT_FOUND_ERROR_MESSAGE));
         authorRepository.deleteById(id);
     }
 
     public void updateAuthor(Long id, AuthorDto authorDto) {
+        if (Objects.isNull(id)||Objects.isNull(authorDto)){
+            throw new BadRequestException(Error.BAD_REQUEST_ERROR_CODE,Error.BAD_REQUEST_ERROR_MESSAGE);
+        }
         Author author = authorRepository.findById(id).orElseThrow(
-                () -> new NotFoundException(Error.BOOK_NOT_FOUND_ERROR_CODE,
-                        Error.BOOK_NOT_FOUND_ERROR_MESSAGE));
+                () -> new NotFoundException(Error.NOT_FOUND_ERROR_CODE,
+                        Error.AUTHOR_NOT_FOUND_ERROR_MESSAGE));
         author.setName(authorDto.getAuthorName());
         author.setBooks(bookRepository.findAllByIdIn(authorDto.getBookIdList()));
         authorRepository.save(author);
